refactor: DF-266 - Move message functions into messaging folder
