+++ .gitignore
@@ -11,3 +11,4 @@ coverage
 .env
 newman-reports
 *.tsbuildinfo
+test/add-message.js
